Implement CR suggestions WiP
