+++ lib/membrane/core/element/diamond_detection_controller/diamond_detection_state.ex
@@ -1,0 +1,19 @@
+defmodule Membrane.Core.Element.DiamondDetectionController.DiamondDatectionState do
+  @moduledoc false
+
+  alias Membrane.Core.Element.DiamondDetectionController.PathInGraph
+
+  defstruct [
+    :serialized_component_path,
+    ref_to_path: %{},
+    trigger_refs: MapSet.new(),
+    postponed?: false
+  ]
+
+  @type t :: %__MODULE__{
+          serialized_component_path: String.t() | nil,
+          ref_to_path: %{optional(reference()) => PathInGraph.t()},
+          trigger_refs: MapSet.t(reference()),
+          postponed?: boolean()
+        }
+end
