+++ lib/membrane/core/element/state.ex
@@ -9,7 +9,7 @@ defmodule Membrane.Core.Element.State do
 
   alias Membrane.{Clock, Element, Pad, Sync}
   alias Membrane.Core.Child.PadModel
-  alias Membrane.Core.Element.DiamondDetectionController.PathInGraph
+  alias Membrane.Core.Element.DiamondDetectionController.DiamondDatectionState
   alias Membrane.Core.Element.EffectiveFlowController
   alias Membrane.Core.Timer
 
@@ -48,12 +48,7 @@ defmodule Membrane.Core.Element.State do
           satisfied_auto_output_pads: MapSet.t(),
           awaiting_auto_input_pads: MapSet.t(),
           resume_delayed_demands_loop_in_mailbox?: boolean(),
-          diamond_detection: %{
-            serialized_component_path: String.t() | nil,
-            ref_to_path: %{optional(reference()) => PathInGraph.t()},
-            trigger_refs: MapSet.t(reference()),
-            postponed?: boolean()
-          }
+          diamond_detection_state: DiamondDatectionState.t()
         }
 
   # READ THIS BEFORE ADDING NEW FIELD!!!
@@ -86,12 +81,7 @@ defmodule Membrane.Core.Element.State do
             handle_demand_loop_counter: 0,
             pads_to_snapshot: MapSet.new(),
             playback_queue: [],
-            diamond_detection: %{
-              serialized_component_path: nil,
-              ref_to_path: %{},
-              trigger_refs: MapSet.new(),
-              postponed?: false
-            },
+            diamond_detection_state: %DiamondDatectionState{},
             pads_data: %{},
             satisfied_auto_output_pads: MapSet.new(),
             awaiting_auto_input_pads: MapSet.new(),
